requirements.txt
boto3==1.26.165
botocore==1.29.165
facenet_pytorch==2.5.2
Flask==2.3.2
Flask_Cors==4.0.0
mtcnn==0.1.1
numpy==1.23.5
opencv_python==4.8.0.74
Pillow==10.0.0
Pillow==10.0.0
pymongo==4.4.0
Requests==2.31.0
tensorflow==2.12.0
torch==2.0.1
torchvision==0.15.2
Werkzeug==2.3.6
gunicorn
astropy
arcface
